cleanup merge artifacts
